+++ src/main/java/sc/senai/controle_de_estoque/entities/DTO/AlterarCompraDTO.java
@@ -12,7 +12,7 @@ public class AlterarCompraDTO {
     private List<Item> item;
     private Fornecedor fornecedor;
     private Projeto projeto;
-    private Long preco;
+    private Float preco;
     private Date dataCompra;
     private Date dataRecebimento;
     private Date dataInvoice;
@@ -49,11 +49,11 @@ public class AlterarCompraDTO {
         this.projeto = projeto;
     }
 
-    public Long getPreco() {
+    public Float getPreco() {
         return preco;
     }
 
-    public void setPreco(Long preco) {
+    public void setPreco(Float preco) {
         this.preco = preco;
     }
